UPBGE: Don't allocate polygons on heap.
Polygons were allocated on heap without any request of this
as they are allocated only at the game conversion.
To simplify the memory managment and cache performance, the polygons
are used as value instead of pointer in RAS_MeshObject::m_polygons.
In the same time the getter functions of RAS_Polygons are qualified
const.