UPBGE: Implement SCA_InputEvent attribute helpers.
Previously the user needed to read the values or queue list
in an input to know if it was active, inactive, activated or
released.
To simplify this and avoid the lookup in list from python
attributes are added in SCA_InputEvent to do the same, they
are:
inactive
active
activated
released

Note that when they are the status of the input from the last
frame so when an input is released then it is both active
and inactive.