UPBGE: Implement SCA_InputEvent attribute helpers.
Previously the user needed to read the values or queue list
in an input to know if it was up, down, pressed or released.
To simplify this and avoid the lookup in list from python
attributes are added in SCA_InputEvent to do the same, they
are:
up
down
pressed
released

Note that when they are the status of the input from the last
frame so when an input is released then it is both up and down.