UPBGE: Fix derived mesh with custom shaders drawing
The first commit which introduced the bug in this case was:
https://github.com/UPBGE/blender/commit/034494a422b1a8fb8c1d69fded4f534e065b80aa

I didn't wanted to restore all the complex system as before to choose the
draw option but just restore the normal drawing mode for derived mÃ¹eshes
with custom shaders.

I don't know however if this is the good solution.

This is related to this issue: https://github.com/UPBGE/blender/issues/577