UPBGE: Restore bind id for VSM shadow.
Previously the shadow bind id returned by shadowBindId was the depth
map in case of simple shadow or the "moment" shadow in case of VSM.
This was recently unified to return only the depth map in the last
shadow refactor but broken some files using custom shader with VSM
"moment" map.
The previous behaviour is restored while waiting for a better API to
access to the two map in case of VSM.

Fix issue #576.