UPBGE: Remove DerivedMesh render.
Previously the DerivedMesh were sometimes rendered using blender code in
RAS_OpenGLRasterizer. This case happened when the user used a modifiers. These
meshes were not converted as other meshes extracted from DerivedMesh and
converted thanks to BL_ConvertMesh.

This behaviour allows to not convert the mesh to game data and be able to draw
it as it but the disadvantages are that all the rendering code contain two
branches: one for game mesh and an other for derived mesh. This systems ends to
complexe checks, performance decrease for game meshes because of extra checks
and duplication of render code using blender data.

To minimize these issues the decision of converting the DerivedMesh to game data
each time needed is took.

The conversion is optimized to use essentially MLoop data, we now use MLoopUV,
MLoopCol and replaced MFace by MLoopTri which gives us already made triangles,
also MFace support will be abandoned in some times. The conversion process look
as follow:

for each MPoly:
   for each MVert in MPoly:
      convert MLoopUV and MLoopCol
      add vertex and check for redundancy
   for each MLoopTri in MPoly:
      add triangle in DA

Because of the usage of MLoopTri it is impossible to know what MLoopTri are
wrapped by a MFace, in consideration the polygons in game are now compound of
only 3 vertices. KX_PolyProxy fakes the 4th vertex all the time now.
In case the mesh is converted using a wire material, lines were added to the
display array and RAS_Polygon was tracking the real triangles, but later it will
be interesting to not work with triangles at mesh level but triangles at DA
level to avoid looking up to the DA of a polygon for every access. This is
ensured by storing triangle indices in DA, these indices are under
m_triangleIndices, the previous indices are renamed m_primitiveIndices.
Later on physics shape creation we will drop the usage of the shared vertex map,
this map is still used but not stored in RAS_MeshObject and just created
temporary during mesh conversion.
The mesh conversion now uses two main functions: BL_ConvertMesh for converting a
DerivedMesh to RAS_MeshObject, BL_ConvertDerivedMeshToArray to convert a
DerivedMesh to a list of DA corresponding to the materials used by the mesh. The
last function is used by the first function and also by BL_ModifierDeformer.
BL_ConvertDerivedMeshToArray request a list of BL_MeshMaterial, this new struct
store a pointer to the material for a DA and its flags. This list is easily
generated by the two callers of BL_ConvertDerivedMeshToArray.

RAS_Polygon was deleted because is goal was replaced by the triangle indices
stored in DAs, but these indices are relative to the DA and not to the whole
mesh object. Do have them absolute, PolygonRange is implemented in
RAS_MeshObject, this new struct contains a pointer to DA targeted and its start
and end absolute indices. When RAS_MeshObject::GetPolygon is called a loop
iterate over all the range and check if the index passed to the function is
included in the current range, if yes it returns a PolygonInfo instantiated with
the triangle indices of the DA of the range. PolygonInfo contains information
about the DA of the polygon, its vertex indices and its flags inherited from its
material.
Though this system is used only for user polygon access through the python API.

The mesh layer format is changed to not be dependent on MLoopUV or MLoopCol,
indeed as we do multiple conversion with the same DerivedMesh basis we intent on
copying the layers info instead regenerating them, but mesh conversion from a
modifier will not use the same MLoop pointers as it's an other DerivedMesh used.
To solve this issue the layers data store only the index and the type of the
layer. A list of MLoopUV and MLoopCol is generated in
BL_ConvertDerivedMeshToArray matching the layer indices.

As DerivedMesh and RAS_Polygon were removed, the physics shape conversion have
to be rewritten to support DA triangle indices. First of all the physics shape
conversion must use only DA, these DA are get from the mesh object or from the
deformer thanks to GetDisplayArray(index).
The converter iterates over all the DA, get the parent material and check if
this material handles collisions, then a list of the valid DA is generated.
Using this DA list we compute the index and vertex count and also the maximum
vertex original index. This last index is computed inside DA each time a vertex
info is added. With this maximum original vertex index a remapping list is
created and is used to convert vertex position only once per original index.
After the creation of the vertex array, the converter iterates over all the
triangles of all the DA and absolute triangles indices are generated.
In the same time than the shape conversion, the soft body index mapping is
refactored. Instead storing the soft body node index in RAS_VertexInfo, a list
is generated in CcdPhysicsController: m_softBodyIndices. This list match an
original vertex index to its soft body node index.
Finally as the physics shape are unique only per combination of mesh, deformer
and type: the storage of the shapes and the function FindMesh is rewritten to
take a RAS_Deformer as sub identifier.

VBO were a few affected by the suppression of derived mesh render, the only
point is about the frequent regeneration of the DA content and in particular the
change of its size, to handle the case where the size changed the function
RAS_Storage::UpdateSize have to be called.