UPBGE: Fix cone constraint debug draw.
Previously the limits of cone constraint were set to maximumn value
when the value gave by the user was negative. This caused to have
a difference between min and max angle out of range for float and then
created incorrect behaviour while drawing the constraint as it intented
to draw a billion of lines.
This issue is fixed by keeping the user inputs as for other constraints.