UPBGE: Refactor action time management.
Previously the current scene animation time was passed
to BL_Action and internally we compared with the animation
time at start to deduced the time duration from the begining.
This behaviour was perfectly working but requested one extra
member to store the start time.

With the soon coming of activity culling, we need to suspend
the animation as logic but doing this without correcting the
time bring to a time jump when restoring the animation. To
avoid this we can follow the behaviour of scene suspending
but it will make the code more complex. The solution retained
is to not compare current time with start time but incrementing
the animation time each frame  so that the suspending became
implicit.

In consideration the Update function in BL_Action receive an
new argument called deltatime, this value is added to the
m_localframe after multipling by a factor. Even if we increment
time we still need the current time to avoid double update
while updating animation for shadows or texture renderers.

Note: the frame rate is passed once to KX_Scene::UpdateAnimations
and other sub calls instead getting it through KX_KetsjiEngine
every time in BL_Action.