restore shadows in blenderplayer (not working in embedded)
I also restored world probe and irradiance in embedded but it's with a
weird hack (not freeing probe sldata->probes and related data). So this
needs to be changed.

in embedded I have difficulties to know exactly what to free or not (what
data can be reused from the viewport.

When I free all sldata->probes and related data, probes don't work in
embedded. Same for shadows. As there is no data existing data in
Blenderplayer, there is no problem. It would be cool to free all before
game start in embedded but I don't find an easy way. I'll see later.

Or we do 2 different systems for blenderplayer and embedded but it would
be cool to have the same code working for both