fix tonemap for overlay scenes
The principle is that only the last scene must be tonemapped.

But this was a bit confusing because each scene has its own filtermanager.

When we add an overlay scene, we have to change the data of the other
firltermanager (remove the other filtermanager tonemap filter)

Because of that, the code is difficult to read