UPBGE: Fix font loading path from linked files.
When the fonts are used in a linked blend file they commonly uses
a relative path to this linked file and its path is stored in
id.lib->name. But the issue noticed was that this library name
might be also relative to the main file (the one started in game)
and no correction was proceeded to make it an absolute path.
This down to intenting making the font file path absolute with
a relative blend file path.
To fix this issue we ensure that the library path is absolute by
expanding it with the main blend file path. If this library path
was already absolute then nothing is changed.

Note: checked with linked blend files and libload.