UPBGE: Fix action actuator with property.
Previously the action actuator in property mode was simply playing the
action with start and end to 0 and setting after the frame. But this behaviour
is invalid as a action clamp its current frame between start and end, and here
it always resulted in 0.
To fix this issue we restore the previous behaviour and launch a new action
each time with start and end to the property value.

Fix issue: #590.