UPBGE: Fix shader recompilation after error.
Previously when a filter failed its compilation, it was
impossible to recompile new sources.
This was caused by the fact that in RAS_Shader::LinkProgram
the variable m_error was first checked and if it was true
when we discard shader compilation. But this variable was
set when the shader failed for the first time. At the end
the shader always discarded the next compilation.

To fix this issue the compilation discard is removed.

In the same time the 0 and 1 are replaced by true or false for
boolean variable.

Bug reported by TwisterGE.