UPBGE: Allow creating custom filter without shader source.
Previously the user was obligated to specify a fragment source
when calling filter.addFilter for custom filter. But if the user
want to use a custom shader it was impossible to not failed the first
compilation and then call setSourceList.
To avoid an useless compilation failure, the function addFilter allow
creating a custom shader without receiving any fragment source.
This produces a empty filter unused in rendering while the user doesn't
called setSourceList.