UPBGE: Remove GLSL inverse function for version < 140 on ATI GPU.
ATI drivers seems to define a inverse function even if the GLSL version is
1.30 or under. This can be caused by a wrong version read in GLSL or
drivers providing inverse function before the standard.

At this time none of these hypothesis were able to be tested.
The only fix we can propose is to disable the inverse support for version
under 1.40 on ATI GPUs.

Supposed to fix issues: #556, #553, #523.