UPBGE: Allow to remove item by CValue type in CListValue.
Previously when the CListValue was refactored to support templates
the function RemoveValue as the other functions looked for a item
template type instead of previous CValue type. This behaviour is
fine for adding, inserting, finding element but for removing it
could cause some error.

If we look at the usage of RemoveValue function in KX_Scene when
removing an object we can see that we test if a game object pointer
is hold by a collection of list, if a list hold it we remove the
item and decrease the item reference counter until the item is
really deleted. When the item is finally deleted any usage of its
pointer must be only for finding the item to remove in a list.
This worked fine previously but now as we need to upcast from
CValue or KX_GameObject to specific item type as KX_FontObject or
KX_LightObject to remove an object from respectively font and light
list, and as the upcast need to acces to item information, most
of the time the upcast were invalid.

To fix this issue the function RemoveValue accept to use only
CValue pointer type.