UPBGE: Call ProcessLighting only when detecting custom shader.
Previously ProcessLighting was called in RAS_MeshSlot all the time if
no override shader were used, but only the custom shader are actually
needing this function to update potentially used OpenGL lighting builtins.
In most of the case the update was useless, to avoid spend time on this,
ProcessLighting is called inside KX_BlenderMaterial::ActivateMeshSlot
so that we can detect if we actually use a custom shader.
In the same time it removes the management of node data m_useLighting.