ToneMap rendered texture in ImageRender
In eevee, the image rendered is shaded with node setup :) This is very
cool. This just changes if we compare with old bge because in old
bge the image was not shaded (shadeless)

But we still needed to apply tonemap filter to not have a too much dark
image.

For this I applied the same system as in normal Render using
PostRenderScene

I guess now, mirror, video... work but I didn't tested yet

Coooooooooool

Note: I think we could rename m_offScreen to something more explicit to
tell that this is the offscreen used to render the texture to. I didn't
very well understood m_finalOffScreen/m_offScreen