UPBGE: Implement mesh duplication in python.
The user is now able to duplicate a KX_Mesh using the python function
mesh.copy().

Internally this function copy construct KX_Mesh and RAS_Mesh and its
RAS_MeshMaterial, then the mesh conversion is finalized and registered
to the converter.
For this last operation it was needed to store a KX_Scene pointer into
KX_Mesh and to be able to replace it in case of lib loading.