run DRW_render_game_loop_begin only for the first scene
If we don't want to loose the shading of the first scene, this loop must
not be applied for an overlay scene as DST is freed...

A shader is created as BL_BlenderShader constructor is called.

The material settings of the overlay scene objects are ok,
but the lighting is the one of the first scene. Good enough for now