copy eevee probe codes in ketji + temp work on planars
I don't really know how to "clip" the material on the edges of the
probes. I guess the answer is in eevee_materials searching with the
keyword "clip"

As for Shadows I created a big block inside KetsjiEngine (delimited with
balises), but the code can easily be moved in another file

Notes:
-The obmat is maybe incorrect as we convert the probe as an empty. So
for example, if we take an horizontal planar probe and we press S+15,
after conversion the scale on x and y axis will be good but the scale
on z won't I guess. Don't really know. We have to print the obmat in
eevee and compare
-> so check clipping and obmat