UPBGE: Refactoring scene shader generation.
Previously the shader in a converted scene were loaded in two ways
depending if the scene was libloaded or not.
When the scene wasn't libloaded BL_ConvertMesh managed the shader generation
but in case of libloading this was proceeded in BL_BlenderConverter::MergeScene.

To use a common bahaviour the function BL_BlenderConverter::InitSceneShader is
introduced, this function iterate over all materials and meshes of a given
scene converter and generate both shader and attributes, the function is called
from KX_KetsjiEngine::ConvertScene (new function to convert scene), ConvertMeshSpecial,
LinkBlendFile and MergeAsyncLoads.

In the same time KX_LibLoadStatus is modified to use a list of blender scenes
instead of a pointer to opaque data.