Unify FrameBuffers creation, refactor hdr, restore post processing
1) Unify FrameBuffers creation:
- To make management of all FrameBuffers
in all GE code easier, I thought it was a good idea to have all
framebuffers stored in RAS_Rasterizer::FrameBuffers::m_frameBuffers
list. Like that we can manage framebuffers easily (creation/deletion)
- FrameBuffers are used for 2DFilters, in ImageRender, in the eevee
effects manager and in the normal render loop. So all these class need
to access m_rasterizer to use m_rasterizer->GetFrameBuffer function (we
could rename this function...)

2) Refactor HDR: As we deal with framebuffers, I thought we could unify
internal format texture type in all the code according to canvas
hdrtype. I don't really know if this is a good idea because it removes a
bit flexibility I guess (we can't have different internal types for
Image Rendered and main render for example) but this simplify the code
and I guess this simplify the comprehension of the code (same format in
FrameBuffers and textures) and maybe the utilisation of HDR (we have
only to switch of hdrtype in render panel and this changes hdrtype in
all the code.

But we can restore it later easily if this is a bottleneck for artists.

3) I did a mistake when I replaced offscreens with framebuffers. In
PostRenderEevee I choosed the depth texture intead of the color texture
as source uniform to apply post process effects. I fixed that too