UPBGE: Fix physics shape of hidden object with modifiers.
Previously when a deformer (modifier, armature, softbody) was used on
an hidden object the deformer was created but contained no display
arrays. As the derived mesh usage in physics shape creation was removed
and that the user can switch of physics shape with an hidden object,
the displays arrays must be available to create a physics shape.

To do so the display arrays of deformers are not instantiated in RAS_MeshObject
when creatin an new mesh user but they are instantiated when initializing
the deformer and duplicating it.

Fix bug: #604.