restore previous offscreen/framebuffer management system
Since I try use framebuffers instead of offscreens, I wanted to refactor
the framebuffers management system in bge code to have all framebuffers
stored in RAS_Rasterizer::FrameBuffers::m_frameBuffers list. The goal
was to have an unique management system.

But I realized that it was not good because some framebuffers need to be
unique (2 unique framebuffer per imagerendered for example) whereas others can
be reused. So The previous management system was more appropriated and I
restore it