UPBGE: Fix mesh free from libload.
When a mesh is deleted it destructs its mesh materials and then
the display array buckets used by the basic mesh. In the free of
the libloaded library (FreeBlendFile) the materials were free before
the meshes, but in the deletion of DAB it unregister the DAB from
the material bucket. This caused an issue as the material remove
also the material bucket and then the DAB used a dangling pointer
to unregister them.

Solution found by youle.
Fix issue: #607.