UPBGE: Fix image tile animation.
Previously the last update time and last frame in Image could
be write to nan in blend file, this caused issue as these values
were not initialized to 0 when loading the file and animation
failed to compare the last time and compute a proper frame.

To fix this issue these two values are initialized to 0 in
direct_link_image.

Fix issue: #601.