UPBGE: Improve animation and deformer update performance.
Previously the animations were using parallelism, each step consisted in:
- Detect if the object is an armature
- In case of armature detect a child with satified condition to allow
	updating armature
- Update the animations through the action manager
- In case of armature update the deformers of all children

The deformers were all calling a lazy update of the armature thanks to
function ApplyPose.

This system is quite correct but has some limitation. First it exists a
lot of different callers for updating armature pose (ApplyPose), this
function is called from deformers or from bone parent relation which
doesn't benefit of parallelism.
Also each step were not of the same execution time depending if the
object was an aramture or not, with many children or none. The load
of threads could suffer of this irregularity.

The changes of this commit target these limitations.

Firstly the animations update is separated in three stages :
Actions, Armatures, Deformers. Each step is procceded with tbb using
parallel_for with a custom task class, each task class constructor aim
to list all objects concerned by the stage, e.g AnimationTask list all
animated objects with an action manager and a non suspended manager.

Actions

AnimationTask filters all the animated objects and keeps only the ones with
non-suspended action manager, then the parallel part for each object
checks if it is an armature, if yes all children are tested. Previously
the children were get from KX_GameObject::GetChildren but this functions
is quite expensive as it constructs recursively a new list, an iteration
over child scene graph nodes is preferred. Finally the action manager
is updated similary to before.
Inside BL_Action::Update to avoid substracting the current time by
the suspended time, the time sent to the action manager is now current
time minus the suspended time, also a boolean named redundant is
passed to the action manager to explicitly notify that the update
is redundant (used only for applying armatures on now visible armatures).
Function IsActionsSuspended and UpdateActionManager are not anymore
calling GetActionManager as the context ensure that the action manager
is existing (only object with an action manager are in the animated object
list), instead they directly use m_actionManager.

Armatures

First of all the scene now registers indepently active armatures in
m_armatureList in AddNodeReplica and from the converter via AddArmature.
ArmatureTask is filtering all the scene active armatures and test
for function NeedApplyPose which basically compare the last pose
update time with the last animation update time. Then the parallel
part calls ApplyPose.

Deformers

RAS_Deformer expose two fundamental functions, NeedUpdate and Update.
The first return a non zero flag when the deformer is tagged for update
and used by DeformerTask constructor to register a deformer. Its
implementation in soft body justs return always 1 and for skin/shape deformer
it checks for the previous armature pose update or for modification of
the original deformer display arrays and returns a combinaison of values
from BL_MeshDeformer::UpdateReason. The second function is used to
actually update the deformer with as argument the reson returned by
NeedUpdate to select what data needed to be udpated.
Most of the deformer are now using a separate loop to compute the AABB
to avoid branches and not waste time when the automatic update is
disabled.

Performance is pretty similar and this system is more flexible over
all.