UPBGE: Use only min and max operation for computing AABB.
Previously the code parts computing AABB were using a condition
to detect the first point computed used to initialize the AABB
dimensions. But this method requiered a test at each points and
also to be able to detect the first point.
To avoid these restrictions the AABB dimension is initialized to
float min for max and float max for min. This make sure that any
point will affect the AABB the first time.