UPBGE: Fix deformer conversion.
THe deformer conversion have some extra constraint, first it had to be placed
before the RAS_MeshUser are created with AddMeshUser, secondly BL_ShapeDeformer
have to check the case when parent object is null.

Also the function BL_SkinDeformer::SetArmature became useless and the construtors
in deformers in case of conversion of non-replicated object are removed because
unused anymore.