UPBGE: Implement python component manager.
Previously the inexistance of a python component manager forced
in KX_Scene to iterate over all the object to ask for a component
update. The implementation of a manager with a registering
mechanism help to avoid this iteration and optimize as much as
possible the case where a scene doesn't update any components.