UPBGE: Use raw std::vector instead of EXP_ListValue for KX_GameObject::GetChildren.
The functions GetChildren and GetChildrenRecursive were not used only for
the python API but also for the animation and duplication process, in these
last cases the creation of a std::vector is cheaper and more easy to use.
For the python API prupose, EXP_ListValue now had a constructor taking
a std::vector to copy items from.