add meshes in invisible layers to eevee cache to create their eevee material at conversion
As we use eevee render to draw meshes (in this branch),
we need to create their materials and display arrays during
conversion. So we add them to eevee cache.

Later, we'll have to remove them from psl->material_pass to not draw
them until they have been added to an active layer