UPBGE: Implement client/server update utilities.
These utilities instore a mechanism where multiple clients are
checking for modifications of a common server, but the client
are not looking at the server at the same time and the server
state is changing.
To ensure that all client know that the server is modified,
at each update the server iterate over all the clients registered
and set a modification state to true in each client if the
update flag from the server match a part of a filter from the
clients.

The client class is CM_UpdateClient, the server CM_UpdateServer,
these both class use a template paramter for the type studied
for update in the perspective to avoid messing clients and
servers of different types.