cleanup
I don't really know if this is a good idea but this simplified the stuff
for me. Else, this doesn't made the code more readable...

The scene shading groups are now packed in shgroups ListBases
corresponding to a DRWPass.

each time we add a ListBase of shgroups in KX_Scene, we just have to
push_back the ListBase in the std::vector which stores the ListBases