cleanup new gameobj api
Finally I used the method of discard geometry (replace geometry with an
unique (SHARED by all discarded objects) single vertex which I created
in the cache to discard objects ended and to hide the objects we don't
want to see in invisible layers. This maybe sounds a bit hacky
but I didn't found a better way for now and this seems to work fine.