temp work on culling
when an object is culled using bullet culling test, I use
DiscardGeometry method to transform the object to a single hided vertex.
When the object is not culled anymore I use AddNewGeometryToPasses to
restore geometry.

This works but the number of draw calls increase each time objects are
restored. So when we discard geometry, it would be nice to find a way to
delete 1 DRWCall from pass