UPBGE: Fix parent for physics object conversion.
Previously in CcdPhysicsEnvironment::ConvertObject we iterate over all
the parents from the object tree to find the closer object possibly
used as a compound shape parent. But by doing this we affected the
search of just the root parent object to disable dynamic or just for
tracking (e.g for fh spring).
To found both the root parent and the root compound parent, two
variables are used in COnvertObject, compoundParent and parentRoot.
Also to clairfy, the function in physics controller named SetParentCtrl
is rename SetParentRoot as it was used to set the root parent controller
only, not the parent controller.

Fix issue: #634.