UPBGE: Fix near/radar sensors.
When a near sensor is used a physics controller using a sphere shape
is created through CreateSphereController. But previously the construction
info was memset to 0 instead using the CcdConstructionInfo constructor.
This caused issues by setting the collision mask/group to zero.
The memset is removed as it doesn't do more than the constructor and
that the similar funciton CreateConeController isn't requesting it.

In the same time the usage of memset 0 for PHY_PHY_RayCastResult is
removed in favor of a constructor.