UPBGE: Implement non-callback collision test.
Previously if the user wanted to know if two objects were colliding
he needed a collision sensor or a collision callback and then
check for every objects collided or use a specific mask for target
object only. This technic was pretty complex for an operation looking
simpler from the user point of view.

To know if two object are colliding a method named KX_GameObject.collide(obj)
is implemented in python, this method return a tuple of a boolean
(colliding state) and a list of contact points (KX_CollisionContactPoint)
or None.

This function is using the new internal function named CollisionTest
in the physcis environment which returns a struct with essentially the
same information than the user gets. This function is checking for the
presence of a broadphase pair in bullet cache and then getting the
manifold.